responsive header components
