+++ src/components/header/breadcrum.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts"></script>
 <template>
-  <div class="flex gap-2 text-text px-12 py-5">
+  <div class="flex gap-2 text-text lg:px-12 px-6 py-5">
     <span>Home</span>
     <span class="text-subtitle">/</span>
     <span>Furniture</span>
