+++ src/components/sidebar/index.vue
@@ -4,7 +4,7 @@ import brandFilter from "~/components/sidebar/filters/brand.vue";
 </script>
 
 <template>
-  <div class="flex flex-col gap-20">
+  <div class="hidden md:flex flex-col gap-20">
     <h2 class="px-2 text-text font-bold text-xl">Dining Room Tables</h2>
     <div class="flex flex-col gap-6">
       <category-filter />
